moved build system to flit
