+++ fud/fud/__init__.py
@@ -1,0 +1,3 @@
+"""This is a docstring"""
+
+__version__ = "0.1.0"
