+++ fud/fud/utils.py
